feat: add gender and nationality chart section
